Move ellipsoid parameter retrieval to new QgsEllipsoidUtils class
